New, order independent merge algorithm
This new implementation of the merge modop behaves the same as the old one (*according to the unit tests) - except that it ignores the node order among siblings.

Previous node traversal limitations (especially when not writing the full targeted node structure) should still apply (eg. deep accessing a node will not return to the previously used layer of nodes, but permanently move to the siblings of the deep-accessed node).