Add "explicitMerge" ModOp
This is effectively a more reliable merge with order independent patching (except for items with the same name, they are still targeted in their order relative to each other).

Unlike merge, this ModOp requires the content of the ModOp tag to explicitly mirror the full structure of the targeted node (hence the name). No nodes will be removed or added, only the given pcdata nodes will be replaced.